requirements.txt
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
django-cors-headers==4.3.1
